requirements.txt
asgiref==3.9.1
certifi==2025.7.9
charset-normalizer==3.4.2
dj-database-url==3.0.1
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.10
npm==0.1.1
npx==0.1.6
numpy==2.2.6
openpyxl==3.1.5
optional-django==0.1.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
requests==2.32.4
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
